Lift generic control traits up into a new module.
