+++ src/fixture/profile/colordynamic.rs
@@ -5,7 +5,7 @@ use num_derive::{FromPrimitive, ToPrimitive};
 use strum_macros::{Display as EnumDisplay, EnumIter, EnumString};
 
 use crate::fixture::prelude::*;
-use crate::osc::prelude::*;
+use crate::control::prelude::*;
 
 #[derive(Debug)]
 pub struct Colordynamic {
