+++ src/fixture/profile/radiance.rs
@@ -4,7 +4,7 @@ use anyhow::Result;
 use std::{collections::HashMap, time::Duration};
 
 use crate::fixture::prelude::*;
-use crate::osc::prelude::*;
+use crate::control::prelude::*;
 
 #[derive(Debug)]
 pub struct Radiance {
