+++ src/fixture/profile/wizard_extreme.rs
@@ -3,8 +3,8 @@
 use num_derive::{FromPrimitive, ToPrimitive};
 use strum_macros::{Display as EnumDisplay, EnumIter, EnumString};
 
+use crate::control::prelude::*;
 use crate::fixture::prelude::*;
-use crate::osc::prelude::*;
 
 #[derive(Debug)]
 pub struct WizardExtreme {
